feat: add basic history command (#27)
We now have a basic history implementation. We'll build upon it in
subsequent pull requests. Since we're also aware of background programs
a user may be running in the current sessions, we kill them when we
terminate the session

In the context of this change, I also fixed 2 (!) separate bugs in the
expandPath filesystem function.