feat: basic protobuf support
While this is already a bulky change, it's far from being done.

I started this branch as mean to explore how much work we need to do in
order to support protobuf along with avro. It turns out:

- it's quite some work because of the protobuf nature
- I made it far enough it makes sense we get this in main already

These changes allow for a very basic protobuf support. We only support:

- syntax 3
- scalar types (and not even all of them)

We don't support:

- nested types
- "officially" supported google types (via imports)
- multiple messages in the same proto def

Each of this is enough work to warrant its own commit in the future.

On the other hand, this commit introduced a TestRecord interface and
generalized how tests produce and consume kafka records.