feat: revamp server configuration (#94)
The most notable change is the format. We moved away from a properties
file to a TOML file. While the properties file served us well so far, it
was a quick home-grown solution to get things started. The TOML
configuration is much more readable and flexible. It also paves the way
for the upcoming `typestream.auto.toml` file which is necessary for
"dynamic" configuration.

We also introduced a top-level config project in order to centralize
(and thefore reuse) the configuration logic. While it would fit the
`libs` path, a top level folder seems more appropriate for a project
with `libs` dependencies. The argument isn't too strong though so this
may change in the near future.

fixes #92