fix: mount k8s deployment so the server can find it (#100)
I'm not very happy about mounting in a place and have the server copy
the files into the system config place. But I honestly can't find a
simple way to mount configs into a container (in a r/w directory). As
this may just be my ignorance, it may also change soon